requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
openai==1.3.0
faiss-cpu==1.7.4
pyspark==3.5.0
delta-spark==3.0.0
kafka-python==2.0.2
apache-airflow==2.8.0
boto3==1.34.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.0
requests==2.31.0
pytest==7.4.3
faker==20.1.0
python-dotenv==1.0.0